Fix error in StreamingBuffer::append producing the crash
The crash on the media files playback was due to an error in the
-Wconversion warning removal and setting as size_t a size variable
which needs the sign for the algorithm in the circular buffer

Extra conversions added to explicitly cast precision losses

+++ src/megaapi_impl.cpp
@@ -26818,7 +26818,7 @@ size_t StreamingBuffer::append(const char *buf, size_t len)
     // update the internal state
     size_t currentIndex = inpos;
     inpos += len;
-    size_t remaining = inpos - capacity;
+    int remaining = int(inpos - capacity);
     inpos %= capacity;
     size += len;
     free -= len;
@@ -26830,9 +26830,9 @@ size_t StreamingBuffer::append(const char *buf, size_t len)
     }
     else
     {
-        size_t num = len - remaining;
+        size_t num = size_t(int(len) - remaining);
         memcpy(buffer + currentIndex, buf, num);
-        memcpy(buffer, buf + num, remaining);
+        memcpy(buffer, buf + num, size_t(remaining));
     }
 
     return len;
